requirements/pypi.txt
# These dependencies are installed using:
#
#   pip install --upgrade
#
# This list is the default package list. All pure python packages
# for the production environment at runtime should be listed here.
# Binary (non-pure) packages may also be listed here, but you
# should see if there is a conda package that suits your needs.

click
Django
djangorestframework
django-rest-swagger
unittest-xml-reporting
